include all files in processInfo.files by default
The getExclude() method was excluding things from the
processInfo.files array if it lived in a folder like 'tests/...'

That's perfectly appropriate for excluding for *coverage*, but excluding
form the `files` listing means that we won't be able later to determine
that a test should be re-run when the test file itself or its fixtures
change.

Also, it was forcibly excluding *all* files from @tapjs/test from
receiving coverage, since of course they all live in a folder named
'test'.