find source maps via url, not file path
This was breaking in cases where tap imports a module with mockImport,
because it adds a ?tapmock=... search param.

When looking up a SourceMap via *file* path, this doesn't find anything,
because internally node converts this to a file URL, but of course then
the URL lacks the search param.

The impact of this, besides getting occasionally weird coverage reports
for mock imported files, was that a coverage map would cause everything
to fail, because it was not finding that the mocked module was linked
back to one of the coverage included sources.