newer entries with same externalID clobber
This was causing `tap -n` to keep re-running the same test over and over
again, because it leaves the previous coverage and processinfo data
behind in that case. If the newer entry happens to have an earlier uuid,
then it would keep being overridden by the earlier entry, because
readdir returned it first in the list.