Only attempt final sourcemap lookup one time
If there are several files at the end of the process that all have
source maps that cannot be found, then it would previously perform
multiple calls to `Module.findSourceMap`, which can cause Node to hang.

(See also the initial workaround for this issue 441731c, source maps
preemptively loaded rather than waiting for the end of the process.)

This adds a guard to only run the end-of-process lookup attempt a single
time, which avoids Node 20's mysterious FinalizationRegistry hang.

Fix: #3