do not add loader to NODE_OPTIONS if in args
Otherwise, the order of loaders gets messed up, because NODE_OPTIONS
gets placed *ahead of* the explicit node args.

I seem to recall that this wasn't a problem, or maybe it's just not a
problem unless we're using a loader that has to be placed in a specific
order, like tsimp does.

Note that this cannot be reasonably guaranteed for the
`child_process.exec` and `execSync` methods, because in that case, the
entire command line is one long string, and parsing it in an identical
manner as the shell is far too brittle. But for spawn, fork, and
execFile, where the argv list is an explicit array, we can do the right
thing.