record unrecorded files that show up in coverage
Re: https://github.com/tapjs/tapjs/issues/1004

This is the "defense in depth bandaid" part of the solution to this
problem.

What's happening is that sometimes, a transpiled typescript CommonJS
module will escape the require() hook, and fail to be recorded. Then,
later, when we're collecting coverage, V8's coverage API includes the
file. But, since it doesn't show up the processinfo `files` field, we
assume it must have been excluded intentionally from all PI tracking for
some reason, and so the coverage is discarded.

With this change, any files that are seen for the first time in the
`coverageOnProcessEnd` function, having not already been recorded to
`processInfo.files`, will be checked against the exclusion pattern,
and only skipped if they should in fact be excluded. Otherwise, they are
added to the `files` list, so that they can be properly tracked for
`--watch` and the like.

I am not 100% sure why it is that commonjs transpiled modules files are
excaping being recorded, but I suspect it has something to do with how
the ESM hooks and ts-node's `require.extensions` hook are interacting
with one another. Migration to tsimp may solve the issue, not sure.